chore: move evals to ee folder (#1854)
