+++ web/src/ee/features/evals/components/eval-template-detail.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import Header from "@/src/components/layouts/header";
-import { EvalTemplateForm } from "@/src/features/evals/components/template-form";
+import { EvalTemplateForm } from "@/src/ee/features/evals/components/template-form";
 import { api } from "@/src/utils/api";
 import { type EvalTemplate } from "@langfuse/shared";
 import { useRouter } from "next/router";
