+++ web/src/ee/features/evals/pages/new-configs.tsx
@@ -1,0 +1,26 @@
+import Header from "@/src/components/layouts/header";
+import { EvalConfigForm } from "@/src/ee/features/evals/components/eval-config-form";
+import { api } from "@/src/utils/api";
+
+import { useRouter } from "next/router";
+
+export default function NewConfigsPage() {
+  const router = useRouter();
+  const projectId = router.query.projectId as string;
+
+  const evalTemplates = api.evals.allTemplates.useQuery({
+    projectId,
+    limit: 500,
+    page: 0,
+  });
+
+  return (
+    <div>
+      <Header title="Create eval config" />
+      <EvalConfigForm
+        projectId={projectId}
+        evalTemplates={evalTemplates.data?.templates ?? []}
+      />
+    </div>
+  );
+}
