+++ web/src/pages/project/[projectId]/evals/configs/[configId].tsx
@@ -1,3 +1,3 @@
-import { EvalConfigDetail } from "@/src/features/evals/components/eval-config-detail";
+import { EvalConfigDetail } from "@/src/ee/features/evals/components/eval-config-detail";
 
 export default EvalConfigDetail;
