+++ web/src/server/api/root.ts
@@ -14,7 +14,7 @@ import { observationsRouter } from "@/src/server/api/routers/observations";
 import { sessionRouter } from "@/src/server/api/routers/sessions";
 import { promptRouter } from "@/src/features/prompts/server/prompt-router";
 import { modelRouter } from "@/src/server/api/routers/models";
-import { evalRouter } from "@/src/features/evals/server/router";
+import { evalRouter } from "@/src/ee/features/evals/server/router";
 import { posthogIntegrationRouter } from "@/src/features/posthog-integration/posthog-integration-router";
 import { llmApiKeyRouter } from "@/src/features/llm-api-key/server/router";
